Add files via upload
A simple CPP code which is used to find the cabs at a 50 km proximity using the GPS co-ordinates provided by the person who needs a cab. This has been implemented using Great Circle Distance formula. The GPS details of both person and the cab drivers are stored in the JSON file format. The user id and name of all the cab drivers who are available in 50 km proximity are stored in the answer.json file once the code is compiled. This is how the code works.