Add files via upload
A simple CPP code which is used to find the user-id and name of the nearby cabs using the GPS co-ordinates. This works based on Great Circle Distance formula. The GPS co-ordinates of both person who needs the cab and the driver is stored in a text file JSON format. The user-id and name of the cab drivers is stored int the answers.json file once the code is compiled. This is how the code works.